// src/app/interfaces/otalbanileria.ts
export interface OtAlbanileria{
    id?: number;	
    folio: number;	
    fecha: string;	
    nombre_empleado: string;	
    num_empleado: string;	
    nombre_obra: string;	
    area: string; 	
    turno: string;	
    nombre_personal1: string;	
    nombre_personal2: string;	
    nombre_personal3: string;	
    nombre_personal4: string;	
    nombre_personal5: string;	
    nombre_personal6: string;	
    nombre_personal7: string;	
    nombre_personal8: string;	
    nombre_personal9: string;	
    nombre_personal10: string;	
    nombre_personal11: string;	
    nombre_personal12: string;	
    horario1: string;	
    horario2: string;	
    horario3: string;	
    horario4: string;	
    horario5: string;	
    horario6: string;	
    horario7: string;	
    horario8: string;	
    horario9: string;	
    horario10: string;	
    horario11: string;	
    horario12: string;	
    horariofinal1: string;
    horariofinal2: string;
    horariofinal3: string;
    horariofinal4: string;
    horariofinal5: string;
    horariofinal6: string;
    horariofinal7: string;
    horariofinal8: string;
    horariofinal9: string;
    horariofinal10: string;
    horariofinal11: string;
    horariofinal12: string;
    trabajo_realizado: string;	
    id_material1: string;	
    id_material2: string;	
    id_material3: string;	
    id_material4: string;	
    id_material5: string;	
    id_material6: string;	
    descripcion_material1: string;	
    descripcion_material2: string;	
    descripcion_material3: string;	
    descripcion_material4: string;	
    descripcion_material5: string;	
    descripcion_material6: string;	
    solicitado1: string;	
    solicitado2: string;	
    solicitado3: string;	
    solicitado4: string;	
    solicitado5: string;	
    solicitado6: string;	
    sobrante1: string;	
    sobrante2: string;	
    sobrante3: string;	
    sobrante4: string;	
    sobrante5: string;	
    sobrante6: string;	
    ocupado1: string;	
    ocupado2: string;	
    ocupado3: string;	
    ocupado4: string;	
    ocupado5: string;	
    ocupado6: string;	
    hme1: string;	
    hme2: string;	
    hme3: string;	
    hme4: string;	
    tipo1: string;
    tipo2: string;
    tipo3: string;
    tipo4: string;
    cantidad1: string;	
    cantidad2: string;	
    cantidad3: string;	
    cantidad4: string;	
    medidas: string;	
    avance_dia: string;	
    observaciones: string;	
    fotografia: string;	
    created_at?: string;	
    updated_at?: string;
}